Move BitmapEngine infrastructure into separate file
